+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Fixed
 - Update options without color (#166)
+- Prevent `LastEditedBy` errors on page update (#167)
 
 ## Documentation
 - Document `People` page property (#162)